refactor: move integration-tests to crate
