+++ Cargo.toml
@@ -1,6 +1,11 @@
 [workspace]
 resolver = "2"
-members = ["core", "examples/memory-catalog", "examples/rest-catalog"]
+members = [
+    "core",
+    "examples/memory-catalog",
+    "examples/rest-catalog",
+    "integration-tests",
+]
 
 [workspace.dependencies]
 # Async runtime and utilities
@@ -13,6 +18,7 @@ iceberg-compaction-core = { path = "./core" }
 
 # Data processing and storage
 ctor = "0.2.8"
+datafusion = "45.0.0"
 iceberg = { git = "https://github.com/risingwavelabs/iceberg-rust.git", rev = "a43641762eafe135498b886db2257102f3069484", features = [
     "storage-s3",
     "storage-gcs",
@@ -24,6 +30,8 @@ iceberg-datafusion = { git = "https://github.com/risingwavelabs/iceberg-rust.git
 parquet = { version = "54", features = ["async"] }
 port_scanner = "0.1.5"
 rand = "0.9.0"
+tracing = "0.1"
+uuid = "1.0"
 
 # Error handling
 thiserror = "2.0.11"
