+++ core/Cargo.toml
@@ -8,7 +8,7 @@ async-stream = { workspace = true }
 async-trait = { workspace = true }
 backon = "1.2.0"
 ctor = { workspace = true }
-datafusion = "45.0.0"
+datafusion = { workspace = true }
 derive_builder = "0.20"
 futures = { workspace = true }
 iceberg = { workspace = true }
@@ -36,9 +36,9 @@ sqlx = { version = "0.8.2", default-features = false, features = [
 tempfile = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
-tracing = "0.1"
+tracing = { workspace = true }
 url = { workspace = true }
-uuid = "1.0"
+uuid = { workspace = true }
 
 [lints]
 workspace = true
