+++ integration-tests/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "iceberg-compaction-integration-tests"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+# Main library
+iceberg-compaction-core = { path = "../core" }
+
+# Docker and test utilities
+ctor = { workspace = true }
+port_scanner = { workspace = true }
+tempfile = { workspace = true }
+
+# Async runtime
+async-stream = { workspace = true }
+async-trait = { workspace = true }
+futures = { workspace = true }
+tokio = { workspace = true }
+
+# Iceberg and catalog
+iceberg = { workspace = true }
+iceberg-catalog-memory = { workspace = true }
+iceberg-catalog-rest = { workspace = true }
+iceberg-datafusion = { workspace = true }
+
+# DataFusion and Parquet
+datafusion = { workspace = true }
+parquet = { workspace = true }
+
+# Logging
+tracing = { workspace = true }
+
+# Other utilities
+rand = { workspace = true }
+uuid = { workspace = true }
+
+[lints]
+workspace = true
